+++ frontend/src/__generated/ve/pages/Graphing/GraphingEditor.css.js
@@ -1,22 +1,13 @@
 // src/pages/Graphing/GraphingEditor.css.ts
-var combobox = "gw96js7";
-var comboboxText = "gw96js9";
-var comboboxWrapper = "gw96js8";
 var editGraphHeader = "gw96js0";
 var editGraphPanel = "gw96js1";
-var editGraphSidebar = "gw96js5";
-var editorLabel = "gw96js6";
-var graphBackground = "gw96js4";
-var graphWrapper = "gw96jsd";
-var input = "gw96jsc";
-var menuButton = "gw96jsa";
-var menuButtonInner = "gw96jsb";
+var editGraphSidebar = "gw96js4";
+var editorLabel = "gw96js5";
+var graphBackground = "gw96js3";
+var graphWrapper = "gw96js7";
+var input = "gw96js6";
 var previewWindow = "gw96js2";
-var tagSwitch = "gw96js3";
 export {
-  combobox,
-  comboboxText,
-  comboboxWrapper,
   editGraphHeader,
   editGraphPanel,
   editGraphSidebar,
@@ -24,8 +15,5 @@ export {
   graphBackground,
   graphWrapper,
   input,
-  menuButton,
-  menuButtonInner,
-  previewWindow,
-  tagSwitch
+  previewWindow
 };
