+++ frontend/src/__generated/ve/pages/Graphing/OptionDropdown/styles.css.js
@@ -1,0 +1,7 @@
+// src/pages/Graphing/OptionDropdown/styles.css.ts
+var menuButton = "_4e4xu50";
+var menuButtonInner = "_4e4xu51";
+export {
+  menuButton,
+  menuButtonInner
+};
