+++ frontend/src/__generated/ve/pages/Graphing/Settings/styles.css.js
@@ -1,0 +1,5 @@
+// src/pages/Graphing/Settings/styles.css.ts
+var tagSwitch = "k0as8h0";
+export {
+  tagSwitch
+};
