+++ frontend/src/pages/Graphing/Combobox/styles.css.ts
@@ -1,0 +1,15 @@
+import { style } from '@vanilla-extract/css'
+
+export const combobox = style({
+	width: '100%',
+	height: 28,
+	borderRadius: 6,
+})
+
+export const comboboxWrapper = style({
+	flex: 1,
+})
+
+export const comboboxText = style({
+	lineHeight: '24px',
+})
