+++ frontend/src/pages/Graphing/GraphingEditor.css.ts
@@ -1,5 +1,5 @@
 import { vars } from '@highlight-run/ui/vars'
-import { globalStyle, style } from '@vanilla-extract/css'
+import { style } from '@vanilla-extract/css'
 
 export const editGraphHeader = style({
 	height: 40,
@@ -11,10 +11,6 @@ export const editGraphPanel = style({
 
 export const previewWindow = style({
 	width: 'calc(100% - 320px)',
-})
-
-export const tagSwitch = style({
-	width: '100%',
 })
 
 export const graphBackground = style({
@@ -29,30 +25,6 @@ export const editorLabel = style({
 	marginBottom: 4,
 })
 
-export const combobox = style({
-	width: '100%',
-	height: 28,
-	borderRadius: 6,
-})
-
-export const comboboxWrapper = style({
-	flex: 1,
-})
-
-export const comboboxText = style({
-	lineHeight: '24px',
-})
-
-export const menuButton = style({
-	border: vars.border.divider,
-	width: '100%',
-	flex: 1,
-})
-
-export const menuButtonInner = style({
-	height: '24px',
-})
-
 export const input = style({
 	height: 28,
 })
@@ -63,8 +35,4 @@ export const graphWrapper = style({
 	margin: 'auto',
 	zIndex: 1,
 	backgroundColor: vars.color.white,
-})
-
-globalStyle(`${menuButton} > div`, {
-	width: '100%',
 })
