+++ frontend/src/pages/Graphing/OptionDropdown/styles.css.ts
@@ -1,0 +1,16 @@
+import { vars } from '@highlight-run/ui/vars'
+import { globalStyle, style } from '@vanilla-extract/css'
+
+export const menuButton = style({
+	border: vars.border.divider,
+	width: '100%',
+	flex: 1,
+})
+
+export const menuButtonInner = style({
+	height: '24px',
+})
+
+globalStyle(`${menuButton} > div`, {
+	width: '100%',
+})
