+++ frontend/src/pages/Graphing/Settings/index.tsx
@@ -1,0 +1,3 @@
+export { BarChartSettings } from './BarChartSettings'
+export { LineChartSettings } from './LineChartSettings'
+export { TableSettings } from './TableSettings'
