+++ frontend/src/pages/Graphing/Settings/styles.css.ts
@@ -1,0 +1,5 @@
+import { style } from '@vanilla-extract/css'
+
+export const tagSwitch = style({
+	width: '100%',
+})
